Extract LocalityDetailsWhere
